refactor: extract useSpecsCommandTree
