+++ src/web/src/views/cli/hooks/index.ts
@@ -1,0 +1,8 @@
+export { useSpecsCommandTree } from "./useSpecsCommandTree";
+export type {
+  CLISpecsCommand,
+  CLISpecsHelp,
+  CLISpecsResource,
+  CLISpecsCommandExample,
+  CLISpecsCommandVersion,
+} from "./useSpecsCommandTree";
